remove AnyDeclProtocol
