+++ Sources/MacroToolkit/DeclGroup/Actor.swift
@@ -1,7 +1,7 @@
 import SwiftSyntax
 
 /// Wraps an `actor` declaration.
-public struct Actor: DeclGroupProtocol {
+public struct Actor: DeclGroupProtocol, RepresentableBySyntax {
     /// The underlying syntax node for the `actor` declaration.
     public var _syntax: ActorDeclSyntax
